Rename IWCompat to WOCStdLib and expand it/reorganize it so that it's more compatible.   The new option appears in [Properties | Clang | Language | Use WinObjC standard library].  Defaults to Yes.
We make extensive use of #include_next to override various VC/Windows header files and #undef/#define various things, particularly in with WinSock.  We redefine winsock's timeval to __ws2_timeval, so our definition can be king inside our codebase.  We also #undef min and max in stdio.h.  Going forward, the use of #include_next seems like a reasonably robust solution, though I have concerns about the way we're redefining sockaddr_in to union with sin_len.

In Windows.h, we #undef interface so that it doesn't stomp ObjC's @interface keyword.  To re-enable the COM-style defintiion of interface, #include <COMIncludes.h> prior to including your COM headers, then #include <COMIncludes_End.h>.  This will cause interface to be redefined to struct, and then #undef'd, respectively.  COMIncludes_*.h supports include recursion up do a depth of 10.

The WOCStdLib option has been added to the <ClangCompile> item, so that it can be enabled or disable per source file - if you have a Windows-centric source file to compile, or if WOCStdLib is just too much of a headache, it can be disabled.  When enabled, it adds include/WOCStdLib to the earliest include path, and force-includes WOCStdlib/WOCStdlib.h.