New drop of (right now only Windows 10) projections
The projections now support overriding certain parts of WinRT RuntimeClasses. Right now it is limited to only overriding a single override interface.
There is another commit coming to WOCCatalog showing how to use this, with examples.

To tell the projections a derived class is actually overriding certain behaviour, you must list it on the conforming protocols list. This instructs the creation code to do the complicated glue to route the messages correctly.
Note that you can't override arbitrary functions in WinRT. Only the methods provided by the override protocols may be overriden. These will also only be respected when specified on the @interface declaration of the class.

There have been a lot of improvements to collections as well, including inheriting off of collections.

This is Win10-only for now since it's still not fully tested but I wanted to let people play with it. The Win8.1 projections have not been updated - they will be when I'm convinced things are stable enough. :)