Add "presentationTransform" property to WOCDisplayMode to alter the global orientation of the compositor, and add a picker cell in WOCCatalog->Display Mode to demonstrate its use.
This property can be used to force the application into any of the 4 orientations (portrait, landscapeleft, landscaperight, portrait180).  It should primarily be used on for desktops or other fixed-display devices and allows applications designed for landscape mode to be used more easily.  On desktop builds, the default setting for the presentationTransform property is determined via the UISupportedInterfaceOrientations key in Info.plist.  The first key listed will be the initial value for presentationTransform.  If there is no UISupportedInterfaceOrientations key in Info.plist, it will default to portrait.  The application also has an opportunity to adjust presentationTransform in the +[UIApplication setStartupDisplayMode:] callback, and like all other properties in [UIApplication displayMode], it can be adjusted at any time from the UI thread, provided that -[[UIApplication displayMode] updateDisplaySettings] is called after making any property changes.