Dispatch animation callbacks on the main thread
This requires taking a dependency on libdispatch in StarboardXaml, for the dispatch_async call.
This fixes deadlocks that happened where client code would do a synchronous call into the UI thread on the thread the animation callback came in on. It is unnecessary to block on the animation completion event for this, so queue it to do later on the UI thread, where it won't have to block to do UI things.