CFBridgeBase implementation and unit tests
CFBridgeBase acts as the base class for bridged classes.

It provides a mechanism for allocating such classes (CFBridgeBase::alloc), taking an optional
Objective-C Class to function as the object's type. It is templated on the C version's bridged
analogue, which will include all state for the object.

We use NSAllocateObject's extraSpace parameter to store any private state we want, after the
Objective-C portion of the object (enforced to be only the isa field).

Note that the Objective-C class version can have no ivar state, since this changes the size of
the object and, thus, the position we're storing our private information. We cannot know this
at compile time, so cannot account for the size in the C version's structure.

Conflicts:
	build/Foundation/Foundation.Shared/Foundation.def