Pre-push clang formatter script
This script is based on the pre-commit hook, but is intended to only run
before code review or a push to develop. It does a diff against the
upstream branch, or a specified branch, and runs the clang formatter on
those files.
Some modifications were made to use the newly checked-in clang binary,
as well as enable the script to be run from subdirectories in the repo.

VSO 5243029