Fix NSURLConnection to not leak memory like a sieve.
NSURLConnection used to store the incoming data on itself, even if you
weren't requesting that it keep the data (that is: you wanted a
streaming response). Now, for sync requests, we're storing the data on
the sync fetch state, which is also the connection's delegate.

This kills the vestigial _shouldDiscardData (which was always false :()
and cleans up NSURLConnection a little bit.