Supplement WIL with Objective C exception handling helpers
Some caveats:
This is made quite a bit uglier by the fact that we cannot rethrow excepts in different DLLs than they were originally caught in. To work around this issue, we put the implementation that does all the rethrowing in header files, so it's in the same DLL as the catching code.
Ideally the support code will eventually exist in objcrt's error-hnadling.mm. The rest of the code is in objcrt because it is the lowest level of the stack and everything relies on it. Any support code needs to be accessible everywhere since WIL is exposed everywhere.