Fix issues with git commit trigger
Remove resetting and stashing since it actually does a git clean if -u is
specified and can fail for example when files are locked by VS.

Change pre-commit to fail fast if a file is both staged and unstaged
because the file present on disk does not match the staged candidate to be
committed.

Change pre-push to fail fast if any staged or unstaged changes are present
because what is on the disk may not match what is committed, so the commit
cannot be validated.

Change all scripts to fail fast on command executions with error return
codes, replacing Invoke-Expression (iex) with call operator (&) because
iex cannot capture exit code value and is prone to issues such as breaking
on filenames with spaces.

Consolidate common functions in common.ps1.

VSO 5676266