Enable support for D3D11 Feature Level 9_3
Many Windows Phones do not support full D3D11 and instead only support
feature level 9_3. This commit modifies the OpenGLES code to support 9_3.

It turns out that VS has an OpenGLES project template which has the correct
initialization code required to support the different feature-level
versions, so this is based on that code. This allows us to include WARP
software rendering as well if needed.

Reviewed By: Ramu Ramanathan