Framework improvements for Arm support
This commit removes the dependency on kernel32 from the framework projects
as windows phone does not have kernel32.dll. This change is applied to
every architecture and it allows us to use the VS x86 mobile emulators for
some mobile testing.

This change also fixes the switches used for Arm. (The current switches
picked by the build cause SBC)

Reviewed By: Raj Seshasankaran