Enhance Run-UnitTests.ps1 to log useful error messages on any test module that isn't able to run successfully.
Description:
When a test module was failing to run to completion (due a crash, missing dependency, etc),
the script was spitting out a very not useful error due to the missing results file.
Instead, catch this case and add it to the summary section as a "problem module".
The return code is also adjusted for this to fail builds where this occurs.

Took the opportunity to also add a minimal verbosity option which could be useful
if this script is combined with others / if the spew is too much for some people.

How Verified:
Removed dependencies from Foundation tests and confirmed new output

Reviewed by: duhowett, jejeyara