Enable NSLog debug output in Release mode
Previously, NSLog was piped to the ETL tracing APIs, which do not print to
the debug output in Release mode for performance.
Make an exception for NSLog so app developers can make use of the debug
output in Release mode.

This will be followed by VSO 6240544 to scrub SDK uses of NSLog in favor
of the raw ETL logging APIs.

VSO 6168871
Fixes #307