Enable Islandwood Arm
This change enables support for building arm debug and release.
An x86 and arm compiler and 3rd party libs are included as part of this
change. (Ben Viglietta provided arm libobjc2.)

To compile the framework for arm, some import/exports needed to be fixed.
Specifically, some names are mangled differently on x86 vs arm and so the
Ebr functions were removed from starboard.def and the header files now
properly import/export the functions. This required adding one of the
symbols from PlatformSupport.cpp as a “forced symbol reference” to ensure
that the Ebr functions were pulled into starboard. (As starboard is built
as a lib into itself, the original source files are only pulled in when at
least one thing from the source file is referenced.)

Foundation also needed to be fixed to properly import/export functions.
The current Foundation project was incorrectly importing its own functions
in some cases. Although this is legal, it exposed a bug in the linker.

Also included is changes to the cassowary project:
- add arm support
- remove windows 8.1 support.

The remaining changes are to update the arm project config to better match
with x86 and to use the correct version of icu for arm.

Reviewed By: Dustin Howett and Jordan Saunders