Fix regression in NSRunLoop where an internal method was being used.
Description:
Fix regression in NSRunLoop where an internal method was being used.
This method was removed as part of stubbing from the header so incorrect type information was being used at runtime.
The fix is to change the callers to not use the internal.

How verified:
Sample app on ARM doesn't crash

Reviewed-by: rajsesh, jaredh, benvi