Update projections with safer RTObject interface
The RTObject interface previously varied depending on whether the client
had also included the WRL headers. This led to some unpleasantness for
clients who had not included the headers but wanted to use the parts of
the interface that required them. With this change, nothing in RTObject now
depends on WRL.