TFS: 6548530: 1601 App - Can't click the 'Explore' button because it's below the visible view in the Window
There was a portion of code in UIViewController::setView:(UIView*) that was resizing the view if it exceeded the bounds of the app's window.  In this case, it was taking a 1024x768 view and resizing it to 768x1024 (the app's window size), which resulted in incorrect layout.

I tried setting the same view's size to 1024x1024 on the reference platform, and I reproduced the same layout bug, which leads me to believe that this code is not behaving correctly.

I traced as far back as I could go in our codebase, but I couldn't pinpoint when this code was added, and there are no comments, so it's unclear if removing it will introduce any regressions.

The correct fix appears to be to remove the offending code, and I was unable to find any regressions after doing so.