Rearrange build configuration
This change centralizes configuration settings for the SDK build, so that
they're not duplicated across so many project files.

New files sdk-build.props and sdk-build.targets are introduced, and
framework projects have references to them rather than directly to
sbclang.props and sbclang.targets.

The sdk-build files don't appear in the packaged SDK. Nothing should be
changed for applications, only for building the SDK itself.

This change updates almost all of the SDK projects, except for the unit
tests, which are configured a little differently.