Removing limbo by moving files out to their respective homes
Description:
This change is a good first step in cleaning up the limbo dumping ground of code.
It moves the files to the places they are actually being used currently but does not
make an attempt to fully remove dead code / incorrectly placed code etc. Low hanging fruit
of obviously unused / empty code was removed but further investigation is likely needed to
complete the clean up.

How Verified:
Still builds

Reviewed by: rajsesh, jordansa, duhowett