Fix up some issues with the ARM UT projects stomping preprocessor defines
Description:
Recently ARM UTs were failing and some failing code was temporarily commented out.
A better solution is to disable the tests so that we still get visibility into the problem.
To achieve this, the ARM_DISABLED* macros were made to work across all UT projects and
enhanced to work with test fixtures and parameterized tests.

How verified:
Correctly reported as disabled on arm. (also enhanced run script to report which tests are disabled)

Reviewed by: yiyang, benvi, nithishm, davelamb