Natvis portion of fix for bug #210033: [Islandwood] NSString node duplicated when expanding NSString objects in the locals window
Unfortunately, some additional work was needed on the debugger side to make the changes to the natvis file work.  To keep the natvis
file working with the update 2 debugger, I left the previous NSObject visualizer in place, along with the new one.  The debugger will
automatically fall back to the old one if the new one fails to parse.  (The issue is that "*this,ndexpand(1)" will only parse on update 3,
not update 2).