ARM CoreFoundation fixes
Description:
Previous fix for ARM is again insufficient. It caused __debugbreaks to be emitted
in other uses of __dmb (because it had been redefinied to __emit). This new
fix is more targeted and only solves the ambiguous symbol compilation issue by
replacing it with a #def at the call site.

This change additionally fixes running some UTs on ARM due to a pwd issue and an
issue in the UT script.

How Verified:
UTs + Funcitonal Tests + Sample App on ARM

Reviewed by: jaredh, benvi, rajsesh