Fix issue where PasswordCrediential would return ERROR_NOT_FOUND if password was retrieved without ever having been set.
Description:
Fix issue where PasswordCrediential would return ERROR_NOT_FOUND if password was retrieved without ever having been set.
The fix is wrap the projection object to only get retrieve the password if the wrapper is created from an existing
PasswordCredential (assumption is this is from a query where password must exist) and to only use the getter for
password if the property has been set.

How Verified: UTs and Sample Apps

Reviewed-by: duhowett, jejeyara