Compile with C++14
Change the build to use -std=c++14 instead of -std=c++11

This is needed as that is the minimum standard language supported by the
MSVC CRT/STL

Changing C++14 exposed some issues with one tests:
  - The test was accidently initialization two of the floats of __m128
    instead of initializing the complex
  - “1.19i" is the syntax for UDT constant literals in C++14 which is only
    for std::complex. Switched to using "*I" which works for C and C++
  - Islandwood is missing C11 complex support (now included in this change)

There is still the issue (already existing) that __m128 and float32x4_t are
using MSVC’s implementation which are unions instead of proper vectors and
the initializer list ends up using the “long long” field of float32x4_t
instead of double. I attempted to change float32x4_t to a proper vector,
but then we can’t encode the objective-c type for vector and we get a
warning (this appears to be a known limitation with Clang). For now, I left
them as the MSVC version.