Use the standard clang driver, instead of clang -cc1.
Issues:
LLVM 3.7 has a bug in the include_next behaviour when a header is included
using double quotes. The Visual Studio IDE and the Visual Studio Command
Prompt specify the same Windows Kit include search paths, but in a
different order. The VS IDE order works for our SDK, while the VS Command
Prompt does not. As a result, we are forced to pass VS's $(IncludePath)
directories to clang to enforce the order that we want.

Testing:
Built the SDK; built and ran all samples and unit tests.