Fix App Crash issue due to clipboard.
Description:
Windows clipboard APIs require that the app is "in the foreground" to access
clipboard data. This causes some cases where an app may be launched but minimized,
or not in focus which will cause this to fail.

The quick fix is to simply ignore errors and return nil.

How Verified:
Manually verified equivalent app crash no longer repros.

Reviewed-by: rajsesh