Added CGPathElementInternal to simplify alloc.
CGPath now uses CGPathElementInternal when creating an array of elements.
CGPathElementInternal holds a fixed sized array big enough for the maximum
number of points in an element. CGPathElementInternal inherits from
CGPathElement and adjusts the points pointer to point at its internal points
array as much as possible. Calls to updatePointsArrayPointer() need to be done manually when
allocating CGPathElementInternal arrays dynamically or using memcpy.