Remove the upper limit for font sizes. This may cause a crash when very very large sizes are used, due to insufficient memory. The reason being Xaml doesn't do anything fancy to intersect the dirty region, it just draws the whole thing.
Also, this unmasked an issue with NSLayoutManager getting into a dead
loop. If the glyph could not fit into the container, it was assumed to be
because of some exclusion zone, so it just kept trying forever. Now we
check to see that the returned size is the same as the container (which
means no exclusions were encountered) and draw one glyph.

Remove test for large font sizes, since that's functionally equivalent to
positive font sizes.