Fix the KVO removal-without-context test to not be implementation-dependent.
Previously, it would expect that a widlcard context remove would take
out the first observer, leaving the second. This is not a valid
assumption, as the implementation is free to do as it wishes.

Additionally, this commit wraps some statements that should absolutely
not fail in EXPECT_NO_THROW so that we get proper test logging instead
of generic exception handling.

Ref #703.