Fix our illegal implementation of an observed KVO property.
Since KVO relies heavily on tracking object state, it is illegal for an
observed property getter to return a new object every time it is called.
This breaks the accounting for those objects, as the first result will
be observed, the second will be modified, the third will attempt
deregistration, and so on.

Fixes #703.