Fix a handful of OS X test failures around KVO, KVC, and Keyed Unarchival. (#809)
* Fix the KVO removal-without-context test to not be implementation-dependent.

Previously, it would expect that a widlcard context remove would take
out the first observer, leaving the second. This is not a valid
assumption, as the implementation is free to do as it wishes.

Additionally, this commit wraps some statements that should absolutely
not fail in EXPECT_NO_THROW so that we get proper test logging instead
of generic exception handling.

Ref #703.

* Fix #755 by disabling 50% of our KVC aggregate tests.

* Split NSSecureCoding tests, disable ones where OS X does not comply with docs.

Fixes #702.

* Fix our illegal implementation of an observed KVO property.

Since KVO relies heavily on tracking object state, it is illegal for an
observed property getter to return a new object every time it is called.
This breaks the accounting for those objects, as the first result will
be observed, the second will be modified, the third will attempt
deregistration, and so on.

Fixes #703.