Issue 770: NSThread.Priority tests fail on OSX (#802)
This issue is due to the fact that thread priorities can change. When setting a
thread priority, the system actually decides what the thread's priority will be.
Because of this, using a hard equal in the test will fail. The solution is to
check if the priority is within a certain acceptable margin. This was acceptable
on Windows before since Windows pigeon holes the thread priorities into certain
groups that can be counted on to return as a certain value.

Verified by running unit tests and passing.