Fix NSCalendar tests that fail on OSX Fixes #744
NSCalendar Tests have some failures as well as some unexpected behavior.
Some of the tests actually produce the correct results as described in the
documentation despite OSX actually producing incorrect results. These are
when the nanoseconds of two dates are compared using a calendar, when using
a matching flag that preserves smaller units when matching a leap date such as
February 29th, as well as the range of weeks in months always returning 6
despite not every month having 6 weeks.

These tests have been disabled for OSX.