Fixing NSCharacterSet test failures on OSX (#823)
Fixes #745

There were a couple tests that were just invalid cases, specifically
characterSetWithBitmapRepresentation and characterSetWithContentsOfFile which
are not valid cases when passed nil.

A test was failing due to an incorrect encoding being used for
unicode characters. NSUnicodeStringEncoding appears to default to UTF16
Little Endian encoding as where it defaults to UTF16 Big Endian encoding on
the reference platform. This results in the same data representing two
different characters.

Validated by running unit tests and passing.