NSURL changes for special filepath for Storage Files
Description:
This change adds the ability to create an NSURL with a Storage File and do all the
things one would expect to be able to accomplish with it like reading and writing
NSData and NSStreams. This is accomplished by making use of remnants of a "direct file"
url and the Windows FutureAccessCache to retreive the file at a later point.

EbrFile is refactored, minimized, and then extended to work with Storage Files.

How verified:
New functional test