CoreAnimationTest - test app exercising CALayer APIs and types (#1204)
CHANGES:
=======

We created CoreAnimationTest so we can validate the CALayer refactored design. This test app will be used in an automated fashion through Appium/WinAppDriver shortly. CoreAnimationTest is combination of similar tests from LayerPlayer Swift OSS tutorial app, WOCCatalog and the Obj-C CALayerSubClassTour OSS test app. This checkin contains the both the xcodeproj (iOS) and SLN (UWP) for CoreAnimationTest.

There are several CALayer scenarios covered in this initial version of the app:

     1. Basic properties (hidden, opacity, contents, gravity, etc.)
     2. CAAnimation (basic animation initially, next updates will exercise blocks, keyframes, etc.)
     3. CAShapeLayer (basic shapes, drawing lines, etc)
     4. CAScrollLayer (really does nothing on today's IslandWood implementation)
     5. CAEmitterLayer (pretty but perf is not great on today's IslandWood implementation)

This initial version allows us to validate CoreAnimation changes in lockstep with the new design and refactored code. This checkin also contains an update to Xib2Nib in support of storyboard segues - the code was prefixing a string to support xCode-generated NIB files which was actually blocking UWP consumption of the WinObjC-generated NIB files.