Fixing two newly-found regressions with the compositor refactor.
[News Test App: Main article title is cropped]
The TextBox within UIKit.Label needs to bind its width/height to that of its parent (the containing Grid).

[News Test App: Contents get obscured when scrolling quickly]
Fixed, along with some other pre-existing flip animation issues that existed on develop (the snapshot layer was often the wrong size - larger than the actual contents we're flipping to).  I also fixed another regression with the flip animation in ux/uxe  the center of the flip rotation has to be between 0 and 1, but we were setting it to width/height of the layer divided by 2 (attempting to rotate at the center).  This accidentally worked on develop due to the hard-coded 1x1 CALayerXaml size, but it was incorrect and led to funny fly-in behavior in ux/uxe.  Fixed by removing the calls altogether since the default PlaneProjection CenterOfRotationX/CenterOfRotationY is .5 (the center).