Show modal UI in a XAML popup (#1393)
When Islandwood is running in "library" mode (i.e., ObjC code running as a library inside a larger UWA), any UI presented to the top-level UIWindow will now appear inside a XAML Popup object. This improves upon the previous situation of the UI not appearing anywhere at all because there is no top-level UIWindow and Islandwood doesn't own any part of the viewable XAML hierarchy.

To do this, there's a new thing called _UIPopupViewController that is instantiated as a top-level view controller when Islandwood is initialized in library mode. Its job is just to hold on to that Popup object and make it visible whenever it detects that someone's trying to present top-level UI.