CoreAnimationTest app - updated to handle multiple CALayers for validated UX/Composition changes (#1406)
The CoreAnimationTest app is designed to exercise the CALayer properties and how they affect each other. During refactoring of the WinObjC composition and CA layers, we wanted an app that exercises layers in isolation or against each other.

We also wrote up a test plan to tackle layer properties based on our existing coverage and ones that we felt might be more interesting for regression testing. For the most part, we haven't regressed that much but we've uncovered a deep-rooted transform bug that's been in existence for a while with this tool.

Finally, this tool has an extensible UX for layer management and layer controls. This will serve us well as we expand our test app to cover the remaining CALayer properties.