There was a ~recent regression in our CALayer convertPoint/convertRect implementations, because we switched to point conversion via the backing Xaml elements rather than via the cached CALayer 'priv' state.  The change was for a UIScrollViewer fix.
The problem with this approach is around our CATransaction implementation; we're asking the Xaml element to do point conversion *before* the CATransaction has even been committed, and therefore the Xaml element hasn't always been updated yet.

This change switches to non-batched CATransactions by default, which means that although we still support CATransaction stacks, each *action within a CATransaction is passed to Xaml immediately*.

We need to test and validate this change thoroughly, as it's unclear how this may negatively affect apps.  That being said, in all testing I've done so far, things are working fine, but we should test further before rolling this out in a release.

I added an opt-out mechanism to WOCDisplayMode (which we will deprecate at some point if all goes well), so app devs can switch back to the old behavior if this turns out to be a breaking change for them.

This change means that we are able to integrate even closer with Xaml, and we can start to ditch more of our cached state within CALayer in future changes.  It also means better perf in general, because we can just let Xaml/DComp worry about batching up screen updates - less code is always a good thing.

Note: All apps tested so far appear to function properly with this approach, but our (fragile) implementation of transition animations was broken by this change.  So we're also disabling transition animations (they 'complete' immediately) when in the default non-batched CATransaction mode.

In a subsequent change, we'll remove the old batched CATransaction code, and we'll gut the transition animation code and schedule work to re-implement it properly (it's been the source of several bugs over the past ~year).