Fixing a UIButton backgroundimage-stretching regression found in IslandwoodHouse.  Our UIButton implementation renders its background via the backing CALayer's content image (similar to how this was done in our legacy UIButton implementation). (#1539)
IslandwoodHouse loads its UIButtons from nibs; some of which were setting their contentMode values to UIViewContentModeScaleAspectFit.  We internally translate UIViewContentModeScaleAspectFit on a UIView to kCAGravityResizeAspect on its backing CALayer.  With kCAGravityResizeAspect set on the CALayer, we no longer stretch the button's background image (by design for kCAGravityResizeAspect).

The fix is to *always* use kCAGravityResize for UIButton's CALayer's contentsGravity, so we're sure to always stretch the background image (again, this is how it was done in the legacy UIButton implementation).

Also modifying LayerCoordinator to make it possible to view each layer's ContentsGravity value in the VS debugger.

Filing a separate issue to add samples that exercise this scenario.