Add support+tests for clipping context rendering to an image mask. (#1528)
This pull request adds support for clipping rendering into a context by an image or an image mask in terms of Direct2D opacity brushes.

An image mask is a grayscale image whose pixel values range from 0 to 100%, and those values are taken as inverse alpha. That is: a fully black image (px=0.0) will be rendered through at alpha 1.0; a fully white pixel will be at alpha 0.0 and values between the two will be interpolated appropriately.

Masks can be stacked. This is implemented in terms of a double composite operation; the two opacity brushes are combined and used to draw a black region into a new image. The new image will serve as the final bitmap backing the new opacity brush.

Since Direct2D does not support b/w masks natively, an effort was undertaken to translate them into alpha channel masks.

Black/white masks do, unfortunately, not work with CGBitmapContext as-is. We cannot render an image into a 1-component colorspace bitmap. This requires that this pull request be merged after #1529.

Fixes #1421. Provides a base implementation for #1425.