Address Feedback
Use CTRunTests as an example for how normal GTEST tests can be pseudo nicely translated to TAEF tests. This mechanism is unfortunately a little cumbersome in that it doesn't actually group tests into a single class but it does most closely match the gtest structure. A later review should move the logger and other pieces to a more common location / include mechanism.
This test also shows how TEST_F and TEST_P are translated. More work is needed to handle disabled test metadata and better failure handling but the basics are there so that UTs can be moved over.