Reverting: (#1640)
"Moving the bulk of app initialization off of the app activation path.  We should let the app activate before calling into app code to finish starting up."

It doesn't seem to be needed for perf improvements, and it broke some functional tests.  We'll ressurrect the change if needed, along with fixed tests.

This reverts commit b3e269ee1ddf7bb48f9f4b0f61611d5d381fff10.