Fixing hit testing for CALayers and UIViews that have clear backgrounds (#1685)
* Fixing hit testing for CALayers and UIViews that have clear backgrounds; they should receive pointer/touch input.  Note that this is different than setting the UIView's alpha value to 0.0 (in which case they should not be hit-testable).

The legacy version of the compositor had an 'optimization' for transparent layer backgrounds; if the set background was transparent, it would set the background brush on the corresponding Xaml element to null.  This is incorrect, because Xaml elements with null backgrounds aren't able to receive pointer input, so this breaks hit testing.  I think this used to work pre-compositor-refactor because any layer that didn't have a background would instead have a transparent rectangle for receiving touch input.  Post-refactor, the extra rectangles are no longer needed, so this legacy 'optimization' is incorrect.

Fixes #1479
Fixes #1574

* Adding hit-testing example tests.

* reimport

* s

* Add test for nil background color.

* s

* s